requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
websockets>=12.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.29
aiosqlite>=0.20.0          # SQLite async driver (dev)
asyncpg>=0.29.0            # PostgreSQL async driver (prod)
alembic>=1.13.0

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ── Computer Vision ───────────────────────────────────────────────────────────
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
torch>=2.3.0
torchvision>=0.18.0
Pillow>=10.3.0

# ── Kafka ─────────────────────────────────────────────────────────────────────
confluent-kafka>=2.4.0

# ── Video / Stream ingestion ──────────────────────────────────────────────────
yt-dlp>=2024.1.0            # YouTube and HTTP stream URL resolution

# ── VLM ───────────────────────────────────────────────────────────────────────
transformers>=4.41.0
accelerate>=0.30.0
openai>=1.30.0

# ── MLOps ─────────────────────────────────────────────────────────────────────
mlflow>=2.12.0
dvc[s3]>=3.50.0
kfp>=2.7.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client>=0.20.0
evidently>=0.4.30
python-json-logger>=2.0.7

# ── Tracing ───────────────────────────────────────────────────────────────────
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0

# ── Data ──────────────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
tqdm>=4.66.0

# ── Cloud ─────────────────────────────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0
redis>=5.0.0                # Redis client for rate limiting + caching

# ── Serving ───────────────────────────────────────────────────────────────────
ray[serve]>=2.20.0
